修改斜向的相邻点距离为sqrt(2)，hops改为double类型

+++ src/main/java/com/shenyy/pretendto/pathfactory/algo/AStarAlgo.java
@@ -11,7 +11,7 @@ public class AStarAlgo<T, O> extends PathAlgo {
     private int cells;
     private Node[][] map;
     private int checks = 0;
-    private int length = 0;
+    private double length = 0;
     private int startx = -1;
     private int starty = -1;
     private int finishx = -1;
@@ -47,7 +47,7 @@ public class AStarAlgo<T, O> extends PathAlgo {
                 solving = false;
                 break;
             }
-            int hops = priority.get(0).getHops() + 1;
+            double hops = priority.get(0).getHops() + 1;
             ArrayList<Node> explored = exploreNeighbors(priority.get(0), hops);
             if (explored.size() > 0) {
                 priority.remove(0);
@@ -61,7 +61,7 @@ public class AStarAlgo<T, O> extends PathAlgo {
         }
     }
 
-    public ArrayList<Node> exploreNeighbors(Node current, int hops) {    //EXPLORE NEIGHBORS
+    public ArrayList<Node> exploreNeighbors(Node current, double hops) {    //EXPLORE NEIGHBORS
         ArrayList<Node> explored = new ArrayList<>();    //LIST OF NODES THAT HAVE BEEN EXPLORED
         for (int a = -1; a <= 1; a++) {
             for (int b = -1; b <= 1; b++) {
@@ -79,7 +79,7 @@ public class AStarAlgo<T, O> extends PathAlgo {
         return explored;
     }
 
-    public void explore(Node current, int lastx, int lasty, int hops) {    //EXPLORE A NODE
+    public void explore(Node current, int lastx, int lasty, double hops) {    //EXPLORE A NODE
         if (current.getType() != 0 && current.getType() != 1)    //CHECK THAT THE NODE IS NOT THE START OR FINISH
             current.setType(4);    //SET IT TO EXPLORED
         current.setLastNode(lastx, lasty);    //KEEP TRACK OF THE NODE THAT THIS NODE IS EXPLORED FROM
@@ -91,7 +91,7 @@ public class AStarAlgo<T, O> extends PathAlgo {
         }
     }
 
-    public void backtrack(int lx, int ly, int hops) {    //BACKTRACK
+    public void backtrack(int lx, int ly, double hops) {    //BACKTRACK
         length = hops;
         PathFinding.getInstance().length = length;
         while (hops > 1) {    //BACKTRACK FROM THE END OF THE PATH TO THE START
